requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
channels==4.1.0
channels-redis==4.2.0
celery==5.4.0
redis==5.1.0
django-celery-beat==2.7.0
Pillow==10.4.0
python-decouple==3.8
drf-yasg==1.21.7
